Update to include all emojis present in iOS10
As the list of emojis is growing and XCode is having issues displaying large
emoji strings - loading all known emojis from a file into a set and getting rid
of a string variable in EmojiData.

emoji.txt contains all known emojis but not all variants are listed there -
only some commont ones.

When determining whether a string contains emojis - enumerate it using composed
character sequences and check if each sequence is an emoji. When checking if
composed character sequence is an emoji - first check in the map of known
emojis, if that fails - check by enumerating unicode scalars. For it to be an
emoji - the character sequence needs to contains known emojis, optionally
separated by zero-width joiner, and optionally ending in a variant.

References:
http://emojipedia.org/emoji/
http://unicode.org/emoji/charts/full-emoji-list.html
https://en.wikipedia.org/wiki/Variant_form_(Unicode)